pyproject.toml
[tool.poetry]
name = "cv-match"
version = "0.1.0"
description = "A project using devcontainer and Poetry"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.30.0"
openai = "^1.3.0"
python-dotenv = "^1.0.0"
PyPDF2 = "^3.0.0"
reportlab = "^4.0.0"
# Add your dependencies here
# For example:
# fastapi = "^0.95.0"
# uvicorn = "^0.22.0"
openpyxl = "^3.1.5"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 